Remove a bunch of dead code, and make Mapping versions download from mcpbot export, until MCPBot is updated to properly index on maven.

+++ src/main/java/com/github/parker8283/bon2/util/DownloadUtils.java
@@ -165,6 +165,7 @@ public class DownloadUtils {
             //So delete the output if it exists as it's invalid, and return false
             if (output.exists())
                 output.delete();
+            e.printStackTrace();
         }
 
         return false;
@@ -188,7 +189,7 @@ public class DownloadUtils {
                 read = IOUtils.copy(stream, out, listener);
             }
 
-            if (read != len) {
+            if (read != len && len != -1) {
                 output.delete();
                 throw new IOException("Failed to read all of data from " + con.getURL() + " got " + read + " expected " + len);
             }
